mm: Turn struct contig_page_data into .bss
This patch moves the version of contig_page_data structure from
mm/nobootmem.c from data to .bss section.

As it does not actually reference anything from the __init section,
it is enough to remove the __refdata annotation in order to make
this change.

The benefit of this change is that it reduces the size of .data
by 2144 bytes (as observed using the 'size' command).

Signed-off-by: Iulia Manda <iulia.manda21@gmail.com>
Reviewed-by: Josh Triplett <josh@joshtriplett.org>
Signed-off-by: Josh Triplett <josh@joshtriplett.org>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>